Added try catch and check if arraylists are empty
Added a try catch wrapping the entire method in the menu in case there are literally any unexpected issues so the program doesn't crash entirely. Might do this with the rest of the cases too.

Added if else statement checking both arraylists and printing a message if they are empty